Switched PRIDE projects away from using the REST API due to the inconcistencies between the returned file lists and the FTP server.
