+++ ppx/massive.py
@@ -37,8 +37,6 @@ class MassiveProject(BaseProject):
         """Instantiate a MSVDataset object"""
         super().__init__(msv_id, local, fetch)
         self._url = f"ftp://massive.ucsd.edu/{self.id}"
-        self._parser = FTPParser(self._url)
-        self._metadata = None
 
     def _validate_id(self, identifier):
         """Validate a MassIVE identifier.
@@ -92,50 +90,6 @@ class MassiveProject(BaseProject):
     def description(self):
         """A description of this project."""
         return self.metadata["dataset.comments"]
-
-    def remote_dirs(self, glob=None):
-        """List the project directories in the remote repository.
-
-        Parameters
-        ----------
-        glob : str, optional
-            Use Unix wildcards to return specific files. For example,
-            :code:`"*peak"` would return all directories ending in "peak".
-
-        Returns
-        -------
-        list of str
-            The remote directories available for this project.
-        """
-        dirs = self._parser.dirs
-        if glob is not None:
-            dirs = [d for d in dirs if Path(d).match(glob)]
-
-        return dirs
-
-    def remote_files(self, glob=None):
-        """List the project files in the remote repository.
-
-        Parameters
-        ----------
-        glob : str, optional
-            Use Unix wildcards to return specific files. For example,
-            :code:`"*.mzML"` would return all of the mzML files.
-
-        Returns
-        -------
-        list of str
-            The remote files available for this project.
-        """
-        if self.fetch or self._remote_files is None:
-            self._remote_files = self._parser.files
-
-        if glob is not None:
-            files = [f for f in self._remote_files if Path(f).match(glob)]
-        else:
-            files = self._remote_files
-
-        return files
 
 
 def list_projects():
