+++ ppx/pride.py
@@ -44,9 +44,7 @@ class PrideProject(BaseProject):
     def __init__(self, pride_id, local=None, fetch=False):
         """Instantiate a PrideDataset object"""
         super().__init__(pride_id, local, fetch)
-        self._url = self.rest + self.id
-        self._parser = None
-        self._metadata = None
+        self._rest_url = self.rest + self.id
 
     def _validate_id(self, identifier):
         """Validate a PRIDE identifier.
@@ -68,6 +66,14 @@ class PrideProject(BaseProject):
         return identifier
 
     @property
+    def url(self):
+        """The FTP address associated with this project."""
+        if self._url is None:
+            self._url = self.metadata["_links"]["datasetFtpUrl"]["href"]
+
+        return self._url
+
+    @property
     def metadata(self):
         """The project metadata as a nested dictionary."""
         if self._metadata is None:
@@ -79,7 +85,7 @@ class PrideProject(BaseProject):
                     assert self.fetch
 
                 # Fetch the data from the remote repository
-                self._metadata = get(self.url)
+                self._metadata = get(self._rest_url)
                 with metadata_file.open("w+") as ref:
                     json.dump(self._metadata, ref)
 
@@ -116,67 +122,6 @@ class PrideProject(BaseProject):
     def doi(self):
         """The DOI for this project."""
         return self.metadata["doi"]
-
-    def remote_files(self, glob=None):
-        """List the project files in the remote repository.
-
-        Parameters
-        ----------
-        glob : str, optional
-            Use Unix wildcards to return specific files. For example,
-            :code:`"*.mzML"` would return all of the mzML files.
-
-        Returns
-        -------
-        list of str
-            The remote files available for this project.
-        """
-        if self.fetch or self._remote_files is None:
-            res = get(self.file_rest, params={"accession": self.id})
-            self._remote_files = sorted(
-                [
-                    f["publicFileLocations"][0]["value"].split(
-                        self.id + "/", 1
-                    )[1]
-                    for f in res
-                ]
-            )
-
-        files = self._remote_files
-        if glob is not None:
-            files = [f for f in files if Path(f).match(glob)]
-
-        return files
-
-    def download(self, files, force_=False, silent=False):
-        """Download files from the remote repository.
-
-        These files are downloaded to this project's local data directory
-        (:py:attr:`~ppx.PrideProject.local`). By default, ppx will not
-        redownload files with matching file names already present in the local
-        data directory.
-
-        Parameters
-        ----------
-        files : str or list of str
-            One or more files to be downloaded from the remote repository.
-        force_ : bool, optional
-            Redownload files when files of the of the same name already appear
-            in the local data directory?
-        silent : bool, optional
-            Hide download progress bars?
-
-        Returns
-        -------
-        list of Path objects
-            The paths of the downloaded files.
-
-        """
-        if self._parser is None:
-            ftp_url = self.metadata["_links"]["datasetFtpUrl"]["href"]
-            self._parser = FTPParser(ftp_url)
-
-        return super().download(files=files, force_=force_, silent=silent)
 
 
 def get(url, **kwargs):
